+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # Changelog
 
+## 0.1.2 - WIP
+- pstdev(): Population standard deviation
+- stdev(): Sample standard deviation
+- pvariance(): variance for a population
+- variance(): variance for a sample
+
 ## 0.1.1 - 2022-01-27
 - Create Freq class with static method for managing frequencies table
-- Create Stat class with static methods for basci statistic functions like: mean, mode, median, multimode...
+- Create Stat class with static methods for basic statistic functions like: mean, mode, median, multimode...
 - Refactor Statistics class in order to use logic provided by Freq and Stat class
 - Create ArrUtil with some helpers/functions to manage arrays
 - Add CICD test for PHP 8.1